Remove need for change in IEntityInformationExtractor interface
Instead we build the expression from the string with the property name
