+++ Rhino.Security/Interfaces/IEntityInformationExtractor.cs
@@ -17,12 +17,6 @@ namespace Rhino.Security.Interfaces
 		/// <param name="entity">The entity.</param>
 		Guid GetSecurityKeyFor(TEntity entity);
 
-        /// <summary>
-        /// Get the expression to extract the key for the specified entity
-        /// </summary>
-        /// <returns></returns>
-        Expression<Func<TEntity, Guid>> GetSecurityKeyIdExpression();
-
 		/// <summary>
 		/// Gets the description of this security key, used when rhino security needs
 		/// to generate a message that involves this security key
